molorag_plus/requirements.txt
torch
transformers
peft
qwen_vl_utils
pymupdf
datasets
tqdm
numpy
pillow
